+++ example/lib/main.dart
@@ -126,7 +126,8 @@ class _PinCodeVerificationScreenState extends State<PinCodeVerificationScreen> {
                         fontWeight: FontWeight.bold,
                       ),
                       length: 6,
-                      obscureText: false,
+                      obscureText: true,
+                      obscuringCharacter: '*',
                       animationType: AnimationType.fade,
                       validator: (v) {
                         if (v.length < 3) {